Move CollectionConverters to scala.scalanative.junit.utils
