+++ unit-tests/src/test/scala/java/nio/file/FilesTest.scala
@@ -12,7 +12,7 @@ import org.junit.Test
 import org.junit.Assert._
 
 import scalanative.junit.utils.AssertThrows._
-import scala.scalanative.compat.CollectionConverters._
+import scala.scalanative.junit.utils.CollectionConverters._
 
 class FilesTest {
   import FilesTest._
