+++ unit-tests/src/test/scala/java/util/ArrayDequeTest.scala
@@ -5,7 +5,7 @@ import org.junit.Test
 import org.junit.Assert._
 
 import scala.scalanative.junit.utils.AssertThrows._
-import scala.scalanative.compat.CollectionConverters._
+import scala.scalanative.junit.utils.CollectionConverters._
 
 class ArrayDequeTest {
