+++ unit-tests/src/test/scala/java/util/regex/PatternTest.scala
@@ -8,7 +8,7 @@ import org.junit.Ignore
 import org.junit.Test
 import org.junit.Assert._
 
-import scala.scalanative.compat.CollectionConverters._
+import scala.scalanative.junit.utils.CollectionConverters._
 import scalanative.junit.utils._, AssertThrows._, ThrowsHelper._
 
 class PatternTest {
